requirements.txt
alembic==1.0.8
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.6
cffi==1.12.2
Click==7.0
factory-boy==2.11.1
Faker==1.0.4
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.8
MarkupSafe==1.1.1
marshmallow==2.19.1
more-itertools==6.0.0
pluggy==0.9.0
psycopg2-binary==2.7.7
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pytest==4.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.1
text-unidecode==1.2
Werkzeug==0.15.1
WTForms==2.2.1
